Turned ProtectedRoute into a HOC
